+++ scripts/dzVents/runtime/device-adapters/text_device.lua
@@ -17,7 +17,13 @@ return {
 		device['text'] = device.rawData[1] or ''
 
 		device['updateText'] = function (text)
-			device.update(0, text)
+			--?type=command&param=udevice&idx=30&nvalue=0&svalue=hoihoi
+
+			-- use openurl to trigger followup events
+			local url = domoticz.settings['Domoticz url'] ..
+					'/json.htm?type=command&param=udevice&idx=' ..
+					device.id .. '&nvalue=0&svalue=' .. utils.urlEncode(tostring(text))
+			domoticz.openURL(url)
 		end
 
 	end